feat: viewCount int -> long - chore: image sample url 변경

+++ src/main/java/kr/ontherec/api/modules/post/dto/PostResponseDto.java
@@ -9,8 +9,8 @@ public record PostResponseDto (
         List<String> images,
         String title,
         String content,
-        int viewCount,
-        int likeCount,
+        long viewCount,
+        long likeCount,
         LocalDateTime createdAt,
         LocalDateTime modifiedAt
 ) {}
